renamed stub_data_source package
