+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/stub_data_source/StubRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.igor_shaula.api_polling.data_layer.stub_source
+package com.igor_shaula.api_polling.data_layer.stub_data_source
 
 import com.igor_shaula.api_polling.data_layer.AbstractVehiclesRepository
 import com.igor_shaula.api_polling.data_layer.VehicleDetailsRecord
